# api/openapi.yaml
openapi: 3.1.0
info:
  title: Portwing Agent API
  version: 0.1.0
  description: |
    Portwing is a lightweight remote Docker agent for the Drydock container
    monitoring platform. This specification covers the **Standard Mode** HTTP API
    only (Edge Mode communicates over a WebSocket multiplexed protocol documented
    separately in SPEC.md §3).

    ## Authentication

    All protected endpoints require the agent token, supplied via one of:

    - `Authorization: Bearer <token>` (standard, checked first)
    - `X-Portwing-Token` header
    - `X-Dd-Agent-Secret` header (Drydock agent secret header)

    Token comparison uses `crypto/subtle.ConstantTimeCompare` (timing-safe).
    If no token is configured, authentication is disabled and all requests pass
    through unauthenticated.

    **Rate limiting:** 10 failed attempts per IP per rolling minute. Exceeding
    the threshold returns HTTP 429 until the window expires.

    ## Docker Engine API Proxy

    Any path not matched by a named route is transparently proxied to the local
    Docker daemon (Unix socket). This includes every endpoint in the
    [Docker Engine API](https://docs.docker.com/engine/api/) — streaming
    endpoints (`/logs`, `/attach`, `/exec/*/start`, `/events`, `/build`,
    `/images/create`, `/images/push`), interactive exec hijacking
    (`Upgrade: tcp`), and binary responses. Paths are forwarded verbatim; the
    agent negotiates the API version on startup (fallback: `v1.44`).

...

  contact:
...
servers:
...
components:
  securitySchemes:
    bearerToken:
      type: http
      scheme: bearer
      description: >
        Agent token via Authorization: Bearer. Also accepted in the
        `X-Portwing-Token` header, or Drydock's `X-Dd-Agent-Secret` header.
    portwingToken:
      type: apiKey
      in: header
      name: X-Portwing-Token
      description: Agent token via the X-Portwing-Token header.
    ed25519Signature:
      type: apiKey
      in: header
      name: X-Portwing-Signature
      description: |
        Ed25519 per-request signature (base64url, no padding). When this header
        is present the agent verifies the request via Ed25519 instead of the
        shared token.

        Four headers are required together:

        | Header | Format | Example |
        |--------|--------|---------|
        | `X-Portwing-Key-ID` | hex(SHA-256(raw 32-byte pubkey)[:8]) | `4a3f2b1c9d8e7f6a` |
        | `X-Portwing-Timestamp` | Unix epoch, seconds, decimal | `1749600000` |
        | `X-Portwing-Nonce` | 32 random hex chars (128-bit) | `a1b2c3d4e5f6...` |
        | `X-Portwing-Signature` | base64url (no padding) Ed25519 signature | `d2hh...` |

        The signed canonical string is:
        ```
        METHOD\nPATH\nbody-sha256-hex\nunix-timestamp\nnonce
        ```
        For an empty body, `body-sha256-hex` is the full SHA-256 of the empty
        string: `e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855`.

        On failure, the response includes `X-Portwing-Reason` with one of:
        `timestamp-skew`, `replay`, `unknown-key`, `invalid-signature`.
  schemas:
...
    InfoResponse:
      type: object
      required: [version, dockerVersion, mode, uptime, hostname, agentId, agentName, capabilities]
      properties:
        version:
          type: string
          description: Portwing agent version
          example: "0.1.0"
        dockerVersion:
          type: string
          description: Docker Engine version string
          example: "27.0.3"
        mode:
          type: string
          enum: [standard]
          description: Connection mode (always "standard" in HTTP API)
        uptime:
          type: string
          description: Agent uptime as a Go duration string
          example: "2h34m56.789s"
        hostname:
          type: string
          example: "my-server"
        agentId:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        agentName:
          type: string
          example: "my-server"
        capabilities:
          type: array
          items:
            type: string
          example: ["compose", "exec", "metrics", "events", "dd:watch", "dd:trigger"]
    ComposeRequest:
      type: object
      required: [operation, stackName]
      properties:
        operation:
          type: string
          enum: [up, down, pull, ps, logs, restart, stop, start]
          description: Compose sub-command to execute.
        stackName:
          type: string
          description: >
            Stack name. Used as the subdirectory under `STACKS_DIR`
            when `stackDir` is not supplied.
          example: "myapp"
        stackDir:
          type: string
          description: >
            Explicit subdirectory within `STACKS_DIR`. Overrides the
            `stackName`-derived path. Must not escape the stacks directory
            (path-traversal protected).
          example: "myapp/v2"
        services:
          type: array
          items:
            type: string
          description: >
            Optional list of service names to scope the operation.
            Names starting with `-` are rejected (flag-injection guard).
          example: ["web", "worker"]
        build:
          type: boolean
          description: Pass `--build` to `compose up`.
          default: false
        forceRecreate:
          type: boolean
          description: Pass `--force-recreate` to `compose up`.
          default: false
        removeVolumes:
          type: boolean
          description: Pass `--volumes` to `compose down`.
          default: false
        noDeps:
          type: boolean
          description: Pass `--no-deps` to `compose up`.
          default: false
        tail:
          type: integer
          minimum: 1
          description: Number of log lines to tail (`logs` operation only).
          example: 100
        envVars:
          type: object
          additionalProperties:
            type: string
          description: >
            Environment variables written to `.env.drydock` before execution.
            Keys must match `^[a-zA-Z_][a-zA-Z0-9_]*$`. The following keys
            are denied: `LD_PRELOAD`, `LD_LIBRARY_PATH`, `PATH`,
            `DOCKER_HOST`, `DOCKER_CONFIG`, `DOCKER_CERT_PATH`,
            `DOCKER_TLS_VERIFY`, `DOCKER_CONTEXT`, `HOME`, `SHELL`,
            `BASH_ENV`, `ENV`, `CDPATH`, `IFS`.
          example:
            DATABASE_URL: "postgres://..."
            APP_PORT: "8080"
        files:
          type: object
          additionalProperties:
            type: string
          description: >
            Files to write into the stack directory before execution.
            Keys are paths relative to the stack directory; values are
            file content (plain text) or base64-encoded content prefixed
            with `base64:`. Paths are resolved and verified within the
            stacks directory (path-traversal protected).
          example:
            "docker-compose.yml": "version: '3'\nservices:\n  web:\n    image: nginx\n"
        registryAuth:
          $ref: '#/components/schemas/RegistryAuth'
    RegistryAuth:
      type: object
      required: [server, username, password]
      properties:
        server:
          type: string
          description: Registry hostname (e.g. `ghcr.io`).
          example: "ghcr.io"
        username:
          type: string
          example: "myuser"
        password:
          type: string
          format: password
          description: >
            Registry password or token. Passed via `--password-stdin`
            to `docker login`; never appears in process argument list.
          example: "ghp_..."
    ComposeResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        output:
          type: string
          description: Combined stdout + stderr from the Compose command.
          example: "Container myapp-web-1  Started\n"
        error:
          type: string
          description: Error message if `success` is false.
          example: "exit status 1"
...
paths:
...
  /_portwing/info:
    get:
      operationId: getPortwingInfo
      summary: Agent metadata
      description: >
        Returns agent version, Docker version, connection mode, uptime,
        hostname, and capability list. Bearer auth required.
      tags: [Agent]
      security:
        - bearerToken: []
        - portwingToken: []
      responses:
        "200":
          description: Agent metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InfoResponse'
        "401":
          description: Missing or invalid token.
        "429":
          description: Rate limit exceeded (10 failures/min/IP).
...
  /_portwing/compose:
    post:
      operationId: postPortwingCompose
      summary: Docker Compose operations
      description: >
        Executes a Docker Compose sub-command within the managed stacks
        directory (`STACKS_DIR`, default `/data/stacks`). The request body
        specifies the operation, stack, optional files to write, and optional
        env vars.

        **Security controls applied before execution:**
        - Stack path resolved to absolute and verified within `STACKS_DIR`
          (path-traversal protection).
        - Env var keys validated against `^[a-zA-Z_][a-zA-Z0-9_]*$` and
          checked against the denylist.
        - Service names must not start with `-` (flag-injection prevention).
        - Registry auth passed via `--password-stdin` (never in argv).
      tags: [Compose]
      security:
        - bearerToken: []
        - portwingToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ComposeRequest'
            examples:
              up:
                summary: Bring up a stack
                value:
                  operation: up
                  stackName: myapp
                  build: false
                  forceRecreate: false
              down-with-volumes:
                summary: Remove a stack including volumes
                value:
                  operation: down
                  stackName: myapp
                  removeVolumes: true
              deploy-with-files:
                summary: Write compose file then bring up
                value:
                  operation: up
                  stackName: myapp
                  files:
                    "docker-compose.yml": "version: '3'\nservices:\n  web:\n    image: nginx:latest\n"
                  envVars:
                    APP_PORT: "8080"
      responses:
        "200":
          description: Compose command executed (check `success` field for result).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComposeResponse'
              examples:
                success:
                  value:
                    success: true
                    output: "Container myapp-web-1  Started\n"
                failure:
                  value:
                    success: false
                    output: "Error response from daemon: ..."
                    error: "exit status 1"
        "400":
          description: Invalid or malformed request body.
        "401":
          description: Missing or invalid token.
        "429":
          description: Rate limit exceeded.
        "500":
          description: Internal server error.
...
  /api/portwing/enroll:
    post:
      operationId: enrollPublicKey
      summary: Register an Ed25519 public key (Model C enrollment)
      description: |
        One-shot bootstrap endpoint that registers an Ed25519 public key using
        a pre-configured enrollment token. The token is burned after first
        successful use; subsequent calls are rejected until the agent restarts
        with a new `ENROLLMENT_TOKEN`.

        This endpoint is **only available** when `ENROLLMENT_TOKEN` (and
        `AUTHORIZED_KEYS`) are configured. It is reachable without prior
        authentication — the enrollment token IS the credential — but it is
        rate-limited via the same IP-based rate limiter as auth failures.

        On success the public key is appended to the authorized-keys file and
        the key registry is hot-reloaded. The new key is immediately active.
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enrollment_token, public_key]
              properties:
                enrollment_token:
                  type: string
                  description: The pre-configured enrollment token.
                  example: "a3f2b1c9d8e7f6a4b5c6d7e8f9a0b1c2"
                public_key:
                  type: string
                  description: >
                    Base64 (standard encoding) of the raw 32-byte Ed25519
                    public key to register.
                  example: "AAAAC3NzaC1lZDI1NTE5AAAAIB3..."
      responses:
        "200":
          description: Key enrolled successfully.
          content:
            application/json:
              schema:
                type: object
                required: [key_id]
                properties:
                  key_id:
                    type: string
                    description: hex(SHA-256(pubkey)[:8]) — the Key-ID for subsequent requests.
                    example: "4a3f2b1c9d8e7f6a"
                  comment:
                    type: string
                    description: Comment assigned to the enrolled key.
                    example: "enrolled:4a3f2b1c9d8e7f6a"
        "400":
          description: Invalid request body, bad base64, or wrong key length.
        "401":
          description: Wrong enrollment token, or token already burned.
        "405":
          description: Method not allowed (POST required).
        "429":
          description: Rate limit exceeded.
        "500":
          description: Could not write to the authorized_keys file.
...